Added new cloud functions for system function.
Added cloud functions to reset the device, and push it into safe mode.
Added a "power" on and off function, that clear all the LEDs and stop running the animation while OFF, and just resume the animation when ON.
Added functions to set brightness and saturation on animations, but only set these to be used on the static animations currently.
Added cloud variable for current hue.
Simplified string conversion on publish calls.